requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.1.0
requests==2.32.3
sentence-transformers==3.4.1
faiss-cpu==1.10.0
numpy==2.2.4
pydantic==2.10.6
pydantic-settings==2.8.1
PyPDF==5.4.0
